Move components that are now shared between docs and wiki into a shared folder
