+++ ts/pages/documentation/method_block.tsx
@@ -7,7 +7,7 @@ import {TypeDocNode, Styles, TypeDefinitionByName} from 'ts/types';
 import {utils} from 'ts/utils/utils';
 import {SourceLink} from 'ts/pages/documentation/source_link';
 import {MethodSignature} from 'ts/pages/documentation/method_signature';
-import {AnchorTitle} from 'ts/pages/documentation/anchor_title';
+import {AnchorTitle} from 'ts/pages/shared/anchor_title';
 import {Comment} from 'ts/pages/documentation/comment';
 import {typeDocUtils} from 'ts/utils/typedoc_utils';
